requirements-3.5.txt
absl-py
numpy
gym
imageio
tqdm
pandas
matplotlib
pygame
transforms3d
scikit-image
torch
torchvision
wget
tensorboard
ray[default]
tensorboardX
python-dotenv

